+++ AED_Project/db4odemo/src/Business/VaccinationSession/VaccinationSessionDirectory.java
@@ -38,7 +38,7 @@ public class VaccinationSessionDirectory {
     }
     
     
-    public VaccinationSession addVaccinationSession(String sessionName, String sessionDate,float startTime,float endTime,String sessionManager,String aefiManager, String vaccinator, String vaccineName, double availableDoses, String centerName){
+    public VaccinationSession addVaccinationSession(String sessionName, String sessionDate,float startTime,float endTime,String sessionManager,String aefiManager, String vaccinator, String vaccineName, long availableDoses, String centerName){
         VaccinationSession sm = new VaccinationSession(sessionName, sessionDate,startTime, endTime,sessionManager, aefiManager, vaccinator, vaccineName, availableDoses, centerName);
         vaccinationSessionDirectory.add(sm);
         return sm;
